+++ src/lib/ControlTab.svelte
@@ -48,7 +48,7 @@
   export let changePlaybackSpeedBy: (delta: number) => void;
   export let resetPlaybackSpeed: () => void;
   export let setPlaybackSpeed: (factor: number, autoPlay?: boolean) => void;
-  
+
   export const resetAnimation = undefined as unknown as () => void;
 
   export let shapes: Shape[];