+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.4.1 (2025-12-30)
 
-- Optimized optimization and field logic that caused laggy behavior. allow opening multiple app windows easily. export .pp file. Decouple settings from project files. Better code testing. implement.pp  file association. 
+- Optimized optimization and field logic that caused laggy behavior. allow opening multiple app windows easily. export .pp file. Decouple settings from project files. Better code testing. implement.pp file association.
 
 ## 1.3.5 (2025-12-30)